+++ Encore/src/menus/SongSelectMenu.h
@@ -5,8 +5,6 @@
 #ifndef SONGSELECTMENU_H
 #define SONGSELECTMENU_H
 #include "OvershellMenu.h"
-#include <filesystem>
-#include <map>
 
 class SongSelectMenu : public OvershellMenu {
 public:
@@ -27,7 +25,7 @@ private:
     const float previewPlayDuration = 30.0f;
     const float pauseDuration = 2.5f;
     double phaseStartTime = 0.0;
-    int curSongMenuPos = -1;
+    size_t curSongMenuPos = -1;
     double curTime = 0.0;
     double selectionTime = 0.0;
 };
